requirements.txt
# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.19.0
lxml==5.2.1

# Data processing
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.2

# Cloud storage
boto3==1.34.69

# Geocoding
googlemaps==4.10.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# Logging
loguru==0.7.2

# Rate limiting
ratelimit==2.2.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.12.0

# Development
black==24.3.0
flake8==7.0.0
mypy==1.9.0
isort==5.13.2